requirements.txt
Django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0